Always include node in event callbacks
This is a big change, but I think it would make development a lot more
pleasant.

Similar to working with React in a browser context, there are a lot of
reasons that having a reference to a `Node` is useful.

One of the most basic React DOM cases is text inputs. You create the
input, and use `event.target.value` to set the state of your
component.

In react-blessed, having access to nodes in events would be useful,
but they're largely not available (except `focus` and `blur` events).

Because of this, I end up needing to create refs for almost every
element, which isn't _necessarily_ bad, but my preference would be to
just have access to the node in event callbacks.